lib: Provide option to disable EOB on stop()
For certain use cases of RFNoC blocks the data expected by the block is
only in terms of streams. This is the case by default while using AXI
wrappers' simple mode when the user does not intend to interact with the
headers of the packets and is concerned only with the payload of the
packets.

In such applications when the user is not intending to use the headers,
an additional packet that is sent as the end of burst produces
unexpected results. The situation is worsened upon implementation of
state machines that depend on receiving certain number of packets.

The problem can be resolved in hardware block by consuming/ignoring the
end of burst packet, but that still results in performing operations
that are not relevant to intended designs.

Solution: Instead of always sending the end of burst packet, a parameter
is used to determine whether or not the end of burst packet should be
sent or not.
This parameter has been assigned a default value that performs the
existing functionality of sending the burst packet, while it can be now
configured to not send that packet if the user so wishes.