Update gr-ettus for UHD 4.0
This commit updates gr-ettus to support a subset of RFNoC features
present in UHD 4.0 with GNU Radio 3.8.

**PLEASE NOTE**: UHD 4.0 represents a radical change to the RFNoC
architecture, and as such, this commit represents a radical set of
changes to gr-ettus with many breaking changes, incomplete support for
the in-tree blocks in UHD 4.0, and removed examples and utilities. Over
time, expect support for the missing blocks, examples, and utilities to
be restored incrementally. Your mileage, as they say, may vary.

Currently, only the following RFNoC blocks are supported:

* RX and TX radios
* Digital downconverter (DDC)
* Digital upconverter (DUC)
* RX and TX streamers
* RFNoC graph
* RFNoC FFT block
* RFNoC Fosphor block
* Fosphor display block (if Qt dependencies are present and
`ENABLE_QT=ON` was specified when CMake was invoked)

rfnoc-modtool is not present or supported at this time.

The rfnoc_radio_ddc example demonstrates the use of the RFNoC graph
(device), RX radio, DDC, and receive streamer RFNoC blocks to create a
simple RFNoC graph that downconverts a portion of the spectrum received
by the radio block to baseband, then transfers it into a GNU Radio
flowgraph for further processing.