rfnoc_modtool: Bootstrap for GR 3.8/UHD 4.0
This commit adds a very basic version of rfnoc_modtool that can be used
to bootstrap new out-of-tree RFNoC modules with RFNoC blocks that are
compliant with UHD 4.0 and GNU Radio 3.8. Please note that only a bare
minimum of rfnoc_modtool functionality is currently implemented. Your
mileage, as they say, may vary.

The following rfnoc_modtool operations are supported to some degree:

* newmod, for creating the shell of the out-of-tree module
* add, for adding new RFNoC blocks to the module
* rm, for removing RFNoC blocks from the module

Other rfnoc_modtool operations are currently not supported at this time.

Co-authored-by: Aaron Rossetto <aaron.rossetto@ni.com>
Co-authored-by: Jonathon Pendlum <jonathon.pendlum@ettus.com>