fir filter: Allow building against all UHD versions
This uses UHD_VERSION to determine if UHD's FIR filter block controller
has a multi-channel API or not.
All channel arguments on the GNU Radio block are now zero (from
previously no default).

When using UHD 4.1 or 4.0, and trying to access an invalid channel
index, a bespoke warning is printed to the GR logger.

This also reverts e92e957.